refactoring dir layout and etc
