+++ query/src/catalogs/impls/mod.rs
@@ -13,12 +13,12 @@
 // limitations under the License.
 //
 
+pub use catalog::system_catalog::LOCAL_TBL_ID_BEGIN;
+pub use catalog::system_catalog::SYS_TBL_ID_BEGIN;
+pub use catalog::system_catalog::SYS_TBL_ID_END;
 pub use database_catalog::DatabaseCatalog;
-pub use sub_cats::system_catalog::LOCAL_TBL_ID_BEGIN;
-pub use sub_cats::system_catalog::SYS_TBL_ID_BEGIN;
-pub use sub_cats::system_catalog::SYS_TBL_ID_END;
 
+mod catalog;
 mod database_catalog;
 pub(crate) mod meta_backends;
-mod sub_cats;
 pub(crate) mod util;
