+++ query/src/datasources/mod.rs
@@ -19,7 +19,7 @@ pub(crate) mod dal;
 pub(crate) mod database;
 pub(crate) mod database_engine;
 pub(crate) mod database_engine_registry;
-pub(crate) mod store_client;
 pub(crate) mod table;
 pub(crate) mod table_engine;
 pub(crate) mod table_engine_registry;
+pub(crate) mod util;
