+++ query/src/datasources/table/csv/csv_table.rs
@@ -1,16 +1,17 @@
-// Copyright 2020 Datafuse Labs.
+//  Copyright 2021 Datafuse Labs.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
 
 use std::any::Any;
 use std::fs::File;
@@ -28,7 +29,7 @@ use common_streams::SendableDataBlockStream;
 use crate::catalogs::Table;
 use crate::datasources::common::count_lines;
 use crate::datasources::common::generate_parts;
-use crate::datasources::table::csv_table_stream::CsvTableStream;
+use crate::datasources::table::csv::csv_table_stream::CsvTableStream;
 use crate::sessions::DatabendQueryContextRef;
 
 pub struct CsvTable {
