+++ query/src/datasources/table/csv/csv_table_test.rs
@@ -1,16 +1,17 @@
-// Copyright 2020 Datafuse Labs.
+//  Copyright 2021 Datafuse Labs.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
 
 use std::env;
 
@@ -21,7 +22,7 @@ use common_planners::*;
 use common_runtime::tokio;
 use futures::TryStreamExt;
 
-use crate::datasources::table::csv_table::CsvTable;
+use crate::datasources::table::csv::csv_table::CsvTable;
 
 #[tokio::test]
 async fn test_csv_table() -> Result<()> {
