+++ query/src/datasources/table/remote/remote_table.rs
@@ -1,16 +1,17 @@
-// Copyright 2020 Datafuse Labs.
+//  Copyright 2021 Datafuse Labs.
 //
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
 
 use std::any::Any;
 use std::sync::mpsc::channel;
@@ -31,9 +32,9 @@ use common_store_api::StorageApi;
 use common_streams::SendableDataBlockStream;
 
 use crate::catalogs::Table;
-use crate::datasources::store_client::GetStoreApiClient;
-use crate::datasources::store_client::StoreClientProvider;
-use crate::datasources::table_engine::TableFactory;
+use crate::datasources::table_engine::TableEngine;
+use crate::datasources::util::GetStoreApiClient;
+use crate::datasources::util::StoreClientProvider;
 use crate::sessions::DatabendQueryContextRef;
 
 #[allow(dead_code)]
@@ -194,7 +195,7 @@ impl RemoteTable {
 
 pub struct RemoteTableFactory {}
 
-impl TableFactory for RemoteTableFactory {
+impl TableEngine for RemoteTableFactory {
     fn try_create(
         &self,
         db: String,
