+++ query/src/datasources/util/store_client_provider.rs
